+++ cli/templates/template-ios/xcode-project-ios.py
@@ -29,7 +29,7 @@ def set_cpp_flags_for_files(project, files, flags):
                 build_file.add_compiler_flags(flags)
 
 
-excludes = ["^build$", "^CMakeLists.txt$", "^.*\.md$", "^.*\.js$"]
+excludes = ["^build$", "^CMakeLists.txt$", "^.*\.md$", "^.*\.js$", "^.*\.glsl$"]
 project = XcodeProject.load(f"{proj_ios_name}.xcodeproj/project.pbxproj")
 
 project_target = project.get_target_by_name("template-ios")